WIP address review comments from @tdas
TODO list
* add test which reads checkpoint from 2.3.1 and runs query
